update vendor entity, restock item entity to handle item replenishment history for vendor

+++ server/src/main/java/com/anujsamdariya07/nightshiftInventory/entity/Vendor.java
@@ -47,7 +47,7 @@ public class Vendor {
 
     private int totalRestocks;
 
-    private int totalValue;
+    private double totalValue;
 
     private ArrayList<RestockItem> replenishmentHistory;
